src/catalogservice/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Anoop Rawat <anooprawat109@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.35"
alembic = "^1.7.7"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
